renamed folder landing to lending
