+++ src/app/data/common/hooks/useFetchViewAllUserRewards.ts
@@ -14,7 +14,7 @@ import {
 } from "../../../../shared";
 import { cValueInUsd } from "../math/cValueInUsd";
 import { safetyModuleRewardController } from "@meta";
-import { ViewAllUserRewards } from "../../landing-deprecated/types/ViewAllUserRewards";
+import { ViewAllUserRewards } from "../../lending-deprecated/types/ViewAllUserRewards";
 import { fetchAssetPriceInBlock } from "../queries/AssetPrice.hook";
 import { getConfig } from "../../../utils/queryContractUtils";
