+++ src/app/data/lending-deprecated/hooks/useFetchAssetRewardsData.all.ts
