+++ src/app/data/lending-deprecated/hooks/useFetchUserStrategies.ts
