+++ src/app/data/lending-deprecated/hooks/useFetchViewAssetUtilizationRate.ts
