+++ src/app/data/lending-deprecated/hooks/useFetchViewDetailTotalSupplied.ts
