+++ src/app/data/lending-deprecated/queries/useFetchViewAssetConfiguration.ts
