+++ src/app/data/lending-deprecated/types/ViewApy.ts
