+++ src/app/data/lending-deprecated/types/ViewAssetConfiguration.ts
