+++ src/app/data/lending-deprecated/types/ViewDetailReserveCaps.ts
