+++ src/app/data/lending-deprecated/types/ViewDetailTotalBorrowed.ts
