+++ src/app/data/lending-deprecated/types/ViewDetailTotalSupplied.ts
