+++ src/app/v3/pages/landing-page/tabs/dashboard/portfolio-summary/Summary.tsx
@@ -2,7 +2,7 @@ import { DisplayMoney, FlexCol, Typography } from "@shared";
 import { useFetchFormattedUserProfitAndPortfolio } from "../../../../../../data/ilmv1-deprecated/hooks/user-profit-and-portfolio/UserProfitAndPortfolio.hook";
 import { Profit } from "./Profit";
 import { MorphoUnclaimedRewardsBox } from "./morpho-vaults/rewards/MorphoUnclaimedRewardsBox";
-import { useFetchViewAllUserRewards } from "../../../../../../data/landing-deprecated/hooks/useFetchViewAllRewards";
+import { useFetchViewAllUserRewards } from "../../../../../../data/lending-deprecated/hooks/useFetchViewAllRewards";
 import { UnclaimedRewardsBoxWrapper } from "./UnclaimedRewardsBoxWrapper";
 
 export const PortfolioSummary = () => {
